Refactor logic to determine steps to its own hook
So it's easily testable and also easier to read.
